Dynamic workspace behaviour: add behaviour option that works better with #950 (#956)
This PR adds an option for dynamic workspaces that works much better
with #950 and dynamic workspaces:

#950 now allows defining a space (via winprops) to always open
applications on. With dynamic workspaces, PaperWM will create workspaces
if the set workspace does not exist (which is likely with dynamic
workspaces). This can cause an issue with workspace names as default
dynamic workspaces will remove any empty workspaces.

I've added a PaperWM setting which handles this more consistently and
intuitively:


![image](https://github.com/user-attachments/assets/2990348a-0247-4783-a88c-2a6c527b9f21)